refactor: organize directories and components, create modal context, improve paymaster create form ui
