+++ frontend/components/common/layout/Footer.tsx
