+++ frontend/components/common/layout/LayoutContent.tsx
