+++ frontend/components/common/layout/navbar/Logo.tsx
