+++ frontend/components/home/Description.tsx
@@ -1,0 +1,14 @@
+import { Text } from "@chakra-ui/react";
+import React from "react";
+
+const Description = () => {
+  return (
+    <Text color={"gray.600"} maxW={"2xl"}>
+      Fusion is a simple and easy to use tool to deploy and manage your ZKsync
+      Paymasters. Modularity and simplicity are the core of Fusion. Get started
+      by creating your first Paymaster. It&apos;s free!
+    </Text>
+  );
+};
+
+export default Description;
