+++ frontend/context/index.ts
@@ -1,3 +1,4 @@
 import { AuthProvider, AuthContext } from "./Auth";
+import { ModalProvider, ModalContext } from "./Modal";
 
-export { AuthProvider, AuthContext };
+export { AuthProvider, AuthContext, ModalProvider, ModalContext };
