+++ frontend/hooks/index.ts
@@ -1,3 +1,4 @@
 import useAuth from "./useAuth";
+import useModal from "./useModal";
 
-export { useAuth };
+export { useAuth, useModal };
