+++ frontend/hooks/useModal.ts
@@ -1,0 +1,8 @@
+import { ModalContext } from "@/context";
+import { useContext } from "react";
+
+const useModal = () => {
+  return useContext(ModalContext);
+};
+
+export default useModal;
